Move Timer to core from fn
