+++ reactor-aeron/src/main/java/reactor/aeron/subscriber/AeronSubscriber.java
@@ -26,7 +26,7 @@ import reactor.core.processor.RingBufferProcessor;
 import reactor.core.subscriber.BaseSubscriber;
 import reactor.core.support.ReactiveState;
 import reactor.fn.Consumer;
-import reactor.fn.timer.Timer;
+import reactor.core.timer.Timer;
 import reactor.io.buffer.Buffer;
 
 import java.util.concurrent.atomic.AtomicBoolean;
