+++ reactor-net/src/main/java/reactor/io/net/http/HttpClient.java
@@ -21,7 +21,7 @@ import java.net.InetSocketAddress;
 import org.reactivestreams.Publisher;
 import reactor.Publishers;
 import reactor.fn.Function;
-import reactor.fn.timer.Timer;
+import reactor.core.timer.Timer;
 import reactor.fn.tuple.Tuple2;
 import reactor.io.net.ReactiveChannelHandler;
 import reactor.io.net.ReactiveClient;
