+++ reactor-net/src/main/java/reactor/io/net/impl/netty/http/NettyHttpServer.java
@@ -31,7 +31,7 @@ import org.slf4j.LoggerFactory;
 import reactor.Publishers;
 import reactor.core.error.Exceptions;
 import reactor.core.support.ReactiveState;
-import reactor.fn.timer.Timer;
+import reactor.core.timer.Timer;
 import reactor.io.buffer.Buffer;
 import reactor.io.net.ReactiveChannel;
 import reactor.io.net.ReactiveChannelHandler;
