+++ reactor-net/src/main/java/reactor/io/net/impl/netty/udp/NettyDatagramServer.java
@@ -48,7 +48,7 @@ import org.slf4j.LoggerFactory;
 import reactor.core.error.Exceptions;
 import reactor.core.support.NamedDaemonThreadFactory;
 import reactor.core.support.SignalType;
-import reactor.fn.timer.Timer;
+import reactor.core.timer.Timer;
 import reactor.io.buffer.Buffer;
 import reactor.io.net.ReactiveChannel;
 import reactor.io.net.ReactiveChannelHandler;
