+++ reactor-net/src/main/java/reactor/io/net/tcp/TcpServer.java
@@ -22,7 +22,7 @@ import org.reactivestreams.Publisher;
 import reactor.Processors;
 import reactor.core.support.ReactiveState;
 import reactor.fn.Function;
-import reactor.fn.timer.Timer;
+import reactor.core.timer.Timer;
 import reactor.io.net.Preprocessor;
 import reactor.io.net.ReactiveChannel;
 import reactor.io.net.ReactiveChannelHandler;
